+++ klibs/tests/test_KLExperiment.py
@@ -24,6 +24,7 @@ def run_environment():
     P.ind_vars_file_path = os.path.join(template_path, "independent_variables.py")
     P.ind_vars_file_local_path = os.path.join(template_path, "doesnt_exist.py")
     P.manual_trial_generation = True
+    P.demographics_collected = True
     P.project_name = "PROJECT_NAME"
 
 @pytest.fixture
